+++ packages/fighting-design/_hooks/use-message/index.ts
@@ -120,9 +120,9 @@ const updatePosition = (closeInstance: ComponentInternalInstance): void => {
       instances[placement] as ComponentInternalInstance[]
     )[index]
 
-      /** 减少后面的组件实例偏移量 */
-      ; (instance.exposed as Record<string, Ref<number>>).offsetVal.value -=
-        getNextElementInterval(closeInstance)
+    /** 减少后面的组件实例偏移量 */
+    ;(instance.exposed as Record<string, Ref<number>>).offsetVal.value -=
+      getNextElementInterval(closeInstance)
   }
 }
 
@@ -137,13 +137,13 @@ export const removeInstance = (instance: ComponentInternalInstance): void => {
   /** 当前组件的方位 */
   const placement: MessagePlacement | NotificationPlacement = getPlacement(instance)
 
-    /**
-     * @see Array.prototype.splice() https://developer.mozilla.org/zh-CN/docs/Web/JavaScript/Reference/Global_Objects/Array/splice
-     */
-    ; (instances[placement] as ComponentInternalInstance[]).splice(
-      getIndexByInstance(instance),
-      1
-    )
+  /**
+   * @see Array.prototype.splice() https://developer.mozilla.org/zh-CN/docs/Web/JavaScript/Reference/Global_Objects/Array/splice
+   */
+  ;(instances[placement] as ComponentInternalInstance[]).splice(
+    getIndexByInstance(instance),
+    1
+  )
 }
 
 /**
@@ -191,7 +191,9 @@ export const useMessage = (
     let result: number = options.offset || 20
     /** 获取到当前方位的组件实例集合 */
     const placementInstance: ComponentInternalInstance[] | undefined =
-      instances[options.placement ? options.placement : name === 'message' ? 'top' : 'top-right']
+      instances[
+        options.placement ? options.placement : name === 'message' ? 'top' : 'top-right'
+      ]
 
     if (placementInstance) {
       placementInstance.forEach((instance: ComponentInternalInstance): void => {