+++ packages/fighting-design/_hooks/use-model/index.ts
@@ -10,7 +10,7 @@ export interface UseModelReturn<T, K extends keyof T> {
 
 /**
  * 用于文本框输入类型组件单项数据了丢失问题封装
- * 
+ *
  * @author Tyh2001 <https://github.com/Tyh2001>
  * @param { Object } prop prop 集合
  * @param { string } propName prop 指定的返回键名
@@ -24,7 +24,6 @@ export const useModel = <T, K extends keyof T>(
   emit: Emit,
   emitName: 'update:modelValue'
 ): UseModelReturn<T, K> => {
-
   /**
    * 自定义计算属性
    *