+++ packages/fighting-design/collapse-animation/src/collapse-animation.vue
@@ -41,7 +41,7 @@
    * @param { Object } el 元素节点
    */
   const onBeforeEnter = (el: Element): void => {
-    (el as HTMLElement).style.transition = transitionStyle
+    ;(el as HTMLElement).style.transition = transitionStyle
     ;(el as HTMLElement).style.height = '0'
   }
 
@@ -54,11 +54,11 @@
    */
   const onEnter = (el: Element): void => {
     if ((el as HTMLElement).scrollHeight !== 0) {
-      (el as HTMLElement).style.height = `${(el as HTMLElement).scrollHeight}px`
+      ;(el as HTMLElement).style.height = `${(el as HTMLElement).scrollHeight}px`
     } else {
-      (el as HTMLElement).style.height = ''
+      ;(el as HTMLElement).style.height = ''
     }
-    (el as HTMLElement).style.overflow = 'hidden'
+    ;(el as HTMLElement).style.overflow = 'hidden'
   }
 
   /**
@@ -67,7 +67,7 @@
    * @param { Object } el 元素节点
    */
   const onAfterEnter = (el: Element): void => {
-    (el as HTMLElement).style.transition = ''
+    ;(el as HTMLElement).style.transition = ''
     ;(el as HTMLElement).style.height = ''
   }
 
@@ -79,7 +79,7 @@
    * @param { Object } el 元素节点
    */
   const onBeforeLeave = (el: Element): void => {
-    (el as HTMLElement).style.height = `${(el as HTMLElement).scrollHeight}px`
+    ;(el as HTMLElement).style.height = `${(el as HTMLElement).scrollHeight}px`
   }
 
   /**
@@ -91,7 +91,7 @@
    */
   const onLeave = (el: Element): void => {
     if ((el as HTMLElement).scrollHeight !== 0) {
-      (el as HTMLElement).style.transition = transitionStyle
+      ;(el as HTMLElement).style.transition = transitionStyle
       ;(el as HTMLElement).style.height = '0'
     }
   }
@@ -104,7 +104,7 @@
    * @param { Object } el 元素节点
    */
   const onAfterLeave = (el: Element): void => {
-    (el as HTMLElement).style.transition = ''
+    ;(el as HTMLElement).style.transition = ''
     ;(el as HTMLElement).style.height = ''
   }
 </script>