+++ packages/fighting-design/input/src/props.ts
@@ -60,7 +60,7 @@ export const Props = {
   maxLength: setNumberProp(),
   /**
    * 设置字体大小
-   * 
+   *
    * @see font-size https://developer.mozilla.org/zh-CN/docs/Web/CSS/font-size
    */
   fontSize: setStringNumberProp(),