+++ packages/fighting-design/popup/src/interface.ts
@@ -5,7 +5,7 @@ export type PopupDirection = 'left' | 'right' | 'top' | 'bottom' | 'center'
 
 /**
  * 回调方法类型
- * 
+ *
  * @param { Object } node 元素节点
  */
 export type PopupCallback = (node: Element) => void