+++ packages/fighting-design/time-picker/src/time-picker.vue
@@ -74,7 +74,7 @@
      * 如果非布尔值类型，咋代表点击的是当前时间，则不需要关闭
      */
     if (isBoolean(target)) {
-      (triggerInstance.value as TriggerInstance).handelClose(evt)
+      ;(triggerInstance.value as TriggerInstance).handelClose(evt)
     }
   }
 
@@ -220,9 +220,9 @@
 
         <!-- 操作栏 -->
         <div class="f-time-picker__option">
-          <f-button size="mini" :on-click="evt => onConfirm(evt, 'now')">当前</f-button>
-          <f-button size="mini" :on-click="evt => onConfirm(evt, false)">取消</f-button>
-          <f-button type="primary" size="mini" :on-click="evt => onConfirm(evt, true)">
+          <f-button size="mini" @click="onConfirm($event, 'now')">当前</f-button>
+          <f-button size="mini" @click="onConfirm($event, false)">取消</f-button>
+          <f-button type="primary" size="mini" @click="onConfirm($event, true)">
             确定
           </f-button>
         </div>