+++ start/src/global.d.ts
@@ -97,4 +97,4 @@ declare global {
   const __DEV__: boolean
 }
 
-export { }
+export {}